+++ pkg/sdk3rd/cloudflare/client.go
@@ -91,7 +91,7 @@ func (c *Client) doRequestWithResult(req *resty.Request, res sdkResponse) (*rest
 			return resp, fmt.Errorf("sdkerr: failed to unmarshal response: %w (resp: %s)", err, resp.String())
 		} else {
 			if err := res.GetAPIError(); err != nil {
-				return resp, fmt.Errorf("sdkerr: errors='%s'", err.Error())
+				return resp, fmt.Errorf("sdkerr: api error: %s", err.Error())
 			}
 		}
 	}
