+++ pkg/sdk3rd/ctyun/zz-shared-common/options.go
@@ -1,4 +1,4 @@
-﻿package openapi
+﻿package common
 
 type Options struct {
 	AccessKeyId     string
