+++ pkg/sdk3rd/wangsu/zz-shared-common/client.go
@@ -1,10 +1,9 @@
-package openapi
+package common
 
 import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 	"time"
 
 	"github.com/go-resty/resty/v2"
@@ -16,31 +15,25 @@ type Client struct {
 	rc *resty.Client
 }
 
-func NewClient(baseUrl string, optFns ...OptionsFunc) (*Client, error) {
-	opts := &Options{}
+func NewClient(optFns ...OptionsFunc) (*Client, error) {
+	options := &Options{}
 	for _, fn := range optFns {
-		fn(opts)
+		fn(options)
 	}
 
-	if baseUrl == "" {
-		return nil, fmt.Errorf("sdkerr: unset baseUrl")
+	if options.AccessKey == "" {
+		return nil, fmt.Errorf("sdkerr: unset accessKey")
 	}
-	if _, err := url.Parse(baseUrl); err != nil {
-		return nil, fmt.Errorf("sdkerr: invalid baseUrl: %w", err)
-	}
-	if opts.AccessKeyId == "" {
-		return nil, fmt.Errorf("sdkerr: unset accessKeyId")
-	}
-	if opts.SecretAccessKey == "" {
-		return nil, fmt.Errorf("sdkerr: unset secretAccessKey")
+	if options.SecretKey == "" {
+		return nil, fmt.Errorf("sdkerr: unset secretKey")
 	}
 
 	signer := &signer{
-		accessKeyId:     opts.AccessKeyId,
-		secretAccessKey: opts.SecretAccessKey,
+		accessKey: options.AccessKey,
+		secretKey: options.SecretKey,
 	}
 	httper := resty.New().
-		SetBaseURL(baseUrl).
+		SetBaseURL("https://open.chinanetcenter.com").
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/json").
 		SetHeader("User-Agent", app.AppUserAgent).
