+++ pkg/sdk3rd/wangsu/zz-shared-common/signer.go
@@ -1,4 +1,4 @@
-﻿package openapi
+﻿package common
 
 import (
 	"bytes"
